fix: primitive input 리팩터링
