+++ src/pages/RegisterCard/components/FormInput.tsx
@@ -1,5 +1,5 @@
-import { Input } from "@/components/primitive/Input";
-import { TInputProps } from "@/types/Input";
+import { Input } from "@/components/atom/forms/Input";
+import { TInputProps } from "@/types/input";
 import React, { CSSProperties, useState } from "react";
 import styled from "styled-components";
