--- src/types/Input.tsx
@@ -1,13 +1,0 @@
-import { CSSProperties, InputHTMLAttributes } from "react";
-
-export type TInputProps = Omit<
-  React.DetailedHTMLProps<
-    InputHTMLAttributes<HTMLInputElement>,
-    HTMLInputElement
-  >,
-  "onChange"
-> & {
-  customType?: "textOnly" | "numberOnly";
-  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => string | void;
-  inputStyle?: CSSProperties;
-};
