+++ src/types/input.ts
@@ -1,0 +1,5 @@
+import { ComponentProps } from "react";
+
+export type TInputProps = ComponentProps<"input"> & {
+  variant?: "default" | "borderLess";
+};
